requirements.txt
asgiref==3.11.1
colorama==0.4.6
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
sqlparse==0.5.5
tqdm==4.67.3
tzdata==2025.3
